+++ docs/publish.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+set -eu
+
+cargo publish -p bevy-inspector-egui-derive
+cargo publish -p bevy-inspector-egui
+