+++ docs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## Version 0.25.2
+- introduce features for egui clipboard and open\_url to improve wasm compat
+- re-enable image and mesh inspector impls
+
+## Version 0.25.1
+- don't clamp linearrgba values
+- fix render layer inspector
+- actually make bevy\_egui/render optional
+- put quick.rs plugins in their own schedule to fix system ambiguity
+
+## Version 0.25.0
+- update to bevy 0.14, egui 0.28
+- make `bevy_render` an optional feature
+- put plugin systems in their own Inspect schedule to avoid ambiguities
+
 ## Version 0.24.0
 
 - update to egui 0.27 and bevy_egui 0.27